Initial implementation of a Psychtoolbox installer and a Psychtoolbox updater.
DownloadPsychtoolbox does a checkout of a specific Psychtoolbox-
flavor from the Subversion repository.

UpdatePsychtoolbox performs an update of a PTB working copy
from the Subversion repository.

The Subversion client is needed for this to work.
Only the Psychtoolbox-folder is checked out - This is for
end-users, not for developers!

Initial commit. Needs refinement in the future!