NEW BRANCH/TAG/FLAVOR: Psychtoolbox-3.0.8-PreTiger
The flavor "Psychtoolbox-3.0.8-PreTiger" starts as an
identical copy of the current "beta" before 15.4.2009,
i.e., revision 1245.

This is the state as of the beta release from 18.02.2009.

This is the last release that is (assumed, not tested!)
to work properly in most parts on MacOS/X versions prior
to 10.4 "Tiger", i.e., 10.3 Panther and earlier.

The MEX files of the upcoming beta release are compiled
linked against Apple Frameworks only available for 10.4
and later. Therefore they will likely cease to work on
older systems.

We branch here, so old users have a way to stick to a
known good release for their old and ancient OS/X versions.

This branch will likely never be updated, although updates
for critical bug fixes or enhancements might happen very
infrequently if there is critical demand and the fixes are
easy to integrate.

The PsychtoolboxPostInstallRoutine in the upcoming beta
will output compatibility warnings and informations on how
to switch to this flavor, should it detect a pre 10.4 system.

-mario